Set positional argument correctly
Setting the target and working directory for `crytic-compile` was being done incorrectly. Now, it is done as a function of `config.Target`. Additionally, since `cmd.Dir` is set to `workingDirectory`, we can now search for build artifacts in `BuildDirectory` (or `crytic-export`). There is no reason to do a `filepath.Join()` operation.