Generalized directory creation
Realized that it makes sense to have a `MakeDirectory` function in fs_utils.go instead of a specific one for the corpus.

Additionally, `WriteCorpusToDisk` was changed to create the corpus instead of a specialized function for it.

`ReadCorpusFromDisk` will now safely return if corpus is empty.

This commit just simplifies the directory creation, writing to it, and reading from it process.