feat: support cancun fork (#397)
Upgrades to go-ethereum 1.14.6 to support the Cancun fork, migrates our tracing to the new native tracing, and make our use of the stateDB compatible with the API change. Previously, we were committing the trie for every tx and so we could just restore to that intermediate state and re-execute the tx. Now, the stateDB expects commits to be tied to a block, so I have taken the approach of re-executing the entire tx sequence from the post-deployment state.

---------

Co-authored-by: anishnaik <anish.naik@trailofbits.com>