Fix events collection (#675)
* Fix events collection

GetLogs() walks the `map[common.Hash][]*types.Log` and concatenates
all of the events there. Then it picks the last one. However, the
current tx hash is not necessarily the last one on the concatenation,
which means that the event collected is probably not the desired one.

This changes the code to store the current tx hash OnTxStart and then
use it to query the specific logs for that transaction.

* Fix hash collision for identical transactions from different senders

The transaction object doesn't have the "from" address as that is
implied by the signature on a normal transaction. However, medusa
does not sign transactions. To keep them distinguishable from each
other, this stores the from address in one of the signature fields.